+++ queries/coordination/dolma.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from file_handling import crawl
+from queries.file_handling import crawl
 
 urlpath = Path("./data/dolma_urls.txt")
 querypath = Path("./queries/domains/dolma.sql")
